Added texture support to materials; Added asset loading; Added functionality to flush RenderCommands; Added model texture loading; Moved Streaming to AssetRegistry; Fixed various premake configuration issues; Fixed assimp linking problems; Updated licenses for various dependencies; Updated assimp; Updated stbi; Improved initialization order by properly using the logic thread.

+++ engine/src/graphics/window/glWindowCallbacks.cpp
@@ -49,7 +49,7 @@ namespace emerald {
 				callback(xpos, ypos);
 			}
 
-			EventSystem::dispatch<MousePositionEvent>(xpos, ypos);
+			EventSystem::dispatch<MousePositionEvent>((float)xpos, (float)ypos);
 		});
 
 		glfwSetScrollCallback(window->handle(), [](GLFWwindow* wnd, double xoffset, double yoffset) {
@@ -57,7 +57,7 @@ namespace emerald {
 				callback(xoffset, yoffset);
 			}
 
-			EventSystem::dispatch<MouseScrollEvent>(xoffset, yoffset);
+			EventSystem::dispatch<MouseScrollEvent>((float)xoffset, (float)yoffset);
 		});
 
 		glfwSetCharCallback(window->handle(), [](GLFWwindow* wnd, uint32_t character) {
